REFACTOR: FORMBUILDER COMO COMPONENTE
